Add a new remote procotol that allows the commands to be executed on remote machines.
This involves setting up a little http server, and creating a communication protocol
such that a client can connect to the server, pull a job, execute it, and report back
the status code. We provide a binary, but the user needs to supply a script that will
arrange for these remote executors to be launched on their cluster.

Also add a feature that allows the graph to keep going until a certain number of
failures. Previously it stopped after the first failure, and that's still the default.