Add very simple bearer token auth.
The /start, /begun, /end, and /ping endpoints are designed to only be called from
execgraph-remote, so they required a bearer token to be set in the http headers.
This token is randomly generated when the server starts up and is communicated
to the provisioner via an environment variable. It's transmitted over the network
in clear text. This is not supposed to be real security, it's just to minimize
any confusion that could occur with multiple execgraph running and someone
getting the ports confused.

Also increases the delay an execgraph-remote will wait for work in the case of
ratelimiting to 30 seconds before exiting.

Also expose information about rate limiting on the /status endpoint.

Also improve how soft shutdown is handled.