Allow task to report a surrogate PID.
After sending input to the task on fd4, wait up to 10ms for the task to send
back a u32 on fd3. If this is sent, it will be interpreted as the PID of the
task for purposes of populating the log file, rather than the real PID. The
reason for this is that some workflows use a common proxy task wrapping command
and want to see the PID of the real task rather than the proxy wrapper.